Combined the best of all bots
Except mages
