--- BATTLECODE/My Player/run.bat
@@ -1,2 +1,0 @@
-set PYTHONPATH=..\battlecode\python
-py -3 run.py
