--- botwip/Rocket.py
@@ -1,4 +1,0 @@
-#Rocket.py
-#uhh.....be build
-
-
