--- examplefuncsplayer-c/extra.c
@@ -1,4 +1,0 @@
-// A handy extra function we can call.
-int extra_fuction() {
-    return 27;
-}
