--- examplefuncsplayer-c/extra.h
@@ -1,2 +1,0 @@
-// A handy extra function we can call.
-int extra_function();
