--- examplefuncsplayer-c/run.bat
@@ -1,2 +1,0 @@
-echo C currently doesn't work on Windows without docker. Please use python or run docker.
-exit 1
