--- examplefuncsplayer-java/run.bat
@@ -1,5 +1,0 @@
-rem This file should build and run your code.
-rem It will only run if you're in nodocker mode on a windows machine.
-
-echo Java currently doesn't work on Windows without docker. Please use python or run docker.
-exit 1
