+++ pinkie/run.bat
