+++ pinkie/run.sh
